lab10/hh-back/requirements/base.txt
asgiref==3.8.1
Django==5.1.7
sqlparse==0.5.3
tzdata==2025.1
djangorestframework==3.16.0
django-cors-headers==4.7.0

